Refactored network serialization into OdinMessage classes
Writing to the network has been distributed over many places and classes and it was hard to debug and hard to find bugs because everything had to be written and read in the same order. Refactored into data structures now and you can see all read and write processes in the same class, making it easier to find bugs and to extend the system.